+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.0]
+
+* Improve code style and performance
+* Support zoom using PhotoView
+* Support change background color
+
 ## [0.0.2]
 
 * Upgrade to Dart 2